enations - remove smartheap requirements, clean up some files.
The version of smartheap I added to replace the original version was still severely out of date,
and referred to functions that didn't seem to exist anymore in the version of the CRT I'm linking
against.

This marks the first time that the enations.exe successfully compiles and builds on my computer.