enations - CSpriteParms::CSpriteParms - Allocate full size of sprite header
This fixes a crash when initializing a game. CSpriteHdr contains a flexible
array member at the end of the struct. This is not legal C++, and when
allocating a new CSpriteHdr object, we only allocated enough memory for the
given size of the object, not any additional data. The code immediately
following this allocation attempted to memcpy data into this newly
allocated object. This would then cause a fault as it was attempting to write
past the end of an allocated object.

This patch allocates enough space to avoid blowing up when the memcpy is
called.