enations - area.cpp - Skip sounds if total volume is 0
This avoids a division by zero. When still loading, if CWndArea::UpdateSound()
is called, the size of the window it uses to calculate sound effect volume
will be 0 or -1. This causes the for loops to never run, and thus iTotal is
always 0. If iTotal is 0, the calculations in question will cause a division
by 0.