enations - mem_pool.h - Fix hardcoded size for SMALL_POOL_SIZE
Previously we only allocated 40 bytes (38 rounded to nearest multiple of 4)
when we should have been allocating 48. This is because I had to hardcode the
size of this allocation, since the original smartheap based memory pool
implementation based this value off the sizeof(CMsgBldgStat).

I chose to hardcode this value because I didn't want to create an awkward
dependency cycle between headers.